move cast kind checking to resolver
