+++ main/src/ca/uwaterloo/flix/language/phase/Typer.scala
@@ -903,23 +903,7 @@ object Typer extends Phase[ResolvedAst.Root, TypedAst.Root] {
       case ResolvedAst.Expression.Cast(exp, declaredTyp, declaredEff, tvar, loc) =>
         // A cast expression is unsound; the type system assumes the declared type is correct.
 
-        // type casts must be of star kind
-        def checkTypeCastKind(tpe: Option[Type]): Result[Unit, TypeError] = tpe match {
-          case None => ().toOk
-          case Some(t) if t.kind <:: Kind.Star => ().toOk
-          case Some(t) => TypeError.UninhabitedTypeCast(t, loc).toErr
-        }
-
-        // effect casts must be of bool kind
-        def checkEffectCastKind(tpe: Option[Type]): Result[Unit, TypeError] = tpe match {
-          case None => ().toOk
-          case Some(t) if t.kind <:: Kind.Bool => ().toOk
-          case Some(t) => TypeError.NonBoolEffectCast(t, loc).toErr
-        }
-
         for {
-          _ <- liftRes(checkTypeCastKind(declaredTyp))
-          _ <- liftRes(checkEffectCastKind(declaredEff))
           _ <- liftRes(kindCheckType(declaredTyp.getOrElse(Type.Unit))) // MATT temp hack bc option
           _ <- liftRes(kindCheckType(declaredEff.getOrElse(Type.Unit))) // MATT temp hack bc option
           (actualTyp, actualEff) <- visitExp(exp)
